Disable legacy preview behaviour
The preview panel markup is loaded via LeftAndMain.ss for all
controllers, regardless if they're using preview or not.
The markup has to be available in any subsequent PJAX calls
in order to allow for CMS sections to opt-in to previews.

This opt-in happens by adding a 'cms-previewable' class to the
edit form, which triggers the layout change. As part of the ReactJS
work, the File object implemented the CMSPreviewable interface,
which triggered this class - and subsequently the layout
switching behaviour.

See https://github.com/silverstripe/silverstripe-framework/commit/4be5e7c961f7041dd5babb10d6d5550766f8c7be