Moved AssetGalleryField logic to AssetAdmin
Creating toplevel UIs as forms made sense in PJAX/Entwine for SS3, but React is a more powerful tool where a form schema abstraction can get in the way. The form schema is useful for forms directly relating to records, but should not be the default answer for building UIs like a GridField to show a default list in a ModelAdmin, or a GalleryField to show files in AssetAdmin.

This change makes a "gallery" less portable to other forms, but it'll be relatively easy to wrap the React component into a form field (powered by Entwine), and then drive it off the existing API, or any other API matching the specification. The form field would be a fairly empty shell, with CRUD logic happening elsewhere through the API.

Also, APIs should not be tied to form fields, but rather work towards general purpose APIs for a record type. This means access control to these APIs moves further into the API layer, and is separate from form access rights (or its readonly state).
Forms are still used as intermediary objects to represent form structure and save data with validation constraints enforced, as an intermediary solution until we have a more expressive API system.