FIX UploadField now considers the idea of having cached form state
It is possible for redux-form to cache the state/values of a form now across individual remounts of a form. In this case UploadField overwrites it's own value with that it was given originally. We should assume the value and files are in sync given the files property is not empty