Better formatting of file migration tasks
Don't pretend to generate thumbnails when you're not.
Ideally we could use AssetAdmin->generateThumbnails() for this,
but there's no way to distinguish the "generated" vs. "existing" case
without dropping deep down into ImageManipulation->manipulate().

Since this is a dev task for one off migrations, I think this duplication
is feasible. It means that in the unlikely case where we significantly change the types
of thumbnails generated in AssetAdmin, we might need to adjust this migration task as well.

See https://github.com/silverstripe/silverstripe-framework/issues/8962